requirements.txt
Flask
mysql-connector-python
werkzeug
reportlab==4.0.7